requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.10.2
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.2
sqlparse==0.3.0
